+++ bumpr/__init__.py
@@ -20,16 +20,25 @@ def main():
     from bumpr import log
     log.init()
 
-    from bumpr.config import Config
+    from bumpr.config import Config, ValidationError
     from bumpr.releaser import Releaser
     from bumpr.helpers import BumprError
     from logging import DEBUG, INFO, getLogger
 
     config = Config.parse_args()
     getLogger().setLevel(DEBUG if config.verbose else INFO)
+    logger = getLogger(__name__)
+
+    try:
+        config.validate()
+    except ValidationError as e:
+        msg = 'Invalid configuration: {0}'.format(e)
+        logger.error(msg)
+        sys.exit(1)
+
     try:
         releaser = Releaser(config)
         releaser.release()
     except BumprError as error:
-        getLogger(__name__).error(str(error))
+        logger.error(str(error))
         sys.exit(1)