+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Current
 - Use nosetests instead of custom discovery
 - Some fixes on Python 3 (mostly encodings)
 - Improve error handling
+- Validate configuration
 
 0.2.0 (2013-08-24)
 ------------------